Move PluginSpec.Install to pkg/workspace.InstallPluginCotnent
Move the actual plugin installation code from sdk to pkg. This refactor will allow us to add SDK generation for packages (referenced from PulumiPlugin.yaml) to the plugin installation.

Ref https://github.com/pulumi/pulumi/pull/20524
