+++ changelog/pending/20250922--sdk--move-pluginspec-install-to-pkg-workspace-installplugincotnent.yaml
@@ -1,0 +1,4 @@
+changes:
+- type: chore
+  scope: sdk
+  description: Move PluginSpec.Install to pkg/workspace.InstallPluginContent
