+++ pkg/cmd/pulumi/plugin/plugin_install.go
@@ -266,7 +266,7 @@ func installPluginSpec(
 ) error {
 	// If we got here, actually try to do the download.
 	var source string
-	var payload workspace.PluginContent
+	var payload pkgWorkspace.PluginContent
 	var err error
 	if file == "" {
 		withProgress := func(stream io.ReadCloser, size int64) io.ReadCloser {
@@ -283,7 +283,7 @@ func installPluginSpec(
 		}
 		defer func() { contract.IgnoreError(os.Remove(r.Name())) }()
 
-		payload = workspace.TarPlugin(r)
+		payload = pkgWorkspace.TarPlugin(r)
 	} else {
 		source = file
 		logging.V(1).Infof("%s opening tarball from %s", label, file)
@@ -293,13 +293,13 @@ func installPluginSpec(
 		}
 	}
 	logging.V(1).Infof("%s installing tarball ...", label)
-	if err = install.InstallWithContext(ctx, payload, reinstall); err != nil {
+	if err = pkgWorkspace.InstallPluginContent(ctx, install, payload, reinstall); err != nil {
 		return fmt.Errorf("installing %s from %s: %w", label, source, err)
 	}
 	return nil
 }
 
-func getFilePayload(file string, spec workspace.PluginSpec) (workspace.PluginContent, error) {
+func getFilePayload(file string, spec workspace.PluginSpec) (pkgWorkspace.PluginContent, error) {
 	source := file
 	stat, err := os.Stat(file)
 	if err != nil {
@@ -307,7 +307,7 @@ func getFilePayload(file string, spec workspace.PluginSpec) (workspace.PluginCon
 	}
 
 	if stat.IsDir() {
-		return workspace.DirPlugin(file), nil
+		return pkgWorkspace.DirPlugin(file), nil
 	}
 
 	f, err := os.Open(file)
@@ -328,9 +328,9 @@ func getFilePayload(file string, spec workspace.PluginSpec) (workspace.PluginCon
 		if runtime.GOOS != "windows" && (stat.Mode()&0o100) == 0 {
 			return nil, fmt.Errorf("%s is not executable", source)
 		}
-		return workspace.SingleFilePlugin(f, spec), nil
+		return pkgWorkspace.SingleFilePlugin(f, spec), nil
 	}
-	return workspace.TarPlugin(f), nil
+	return pkgWorkspace.TarPlugin(f), nil
 }
 
 // resolvePluginSpec resolves plugin specifications using various resolution strategies.
