+++ pkg/workspace/plugins_install_nodejs_test.go
