+++ pkg/workspace/plugins_install_python_test.go
